Makes the all books feature more dynamic for reusability.
