+++ src/AllBooks/AllBooks.css
@@ -30,10 +30,3 @@
 .genre-filter {
     margin-top: 5rem;
 }
-/* .to-back {
-    margin: 1rem 0;
-    display: flex;
-    justify-content: center;
-    width: 100%; 
-    z-index: 1;
-} */
