+++ src/MainBook/MainBook.css
@@ -1,7 +1,7 @@
 .main-book-container {
     display:  flex;
     align-items: center; 
-    height: 175vh; 
+    height: 170vh; 
     width: 100vw; 
     flex-direction: column;
   }
